Refactor aggregators.
