+++ akka-http-aggregator/src/main/scala/com/pagerduty/akka/http/aggregator/aggregator/TwoStepAggregator.scala
@@ -1,11 +1,10 @@
 package com.pagerduty.akka.http.aggregator.aggregator
 
 trait TwoStepAggregator[RequestKey, AccumulatedState, AddressingConfig]
-    extends Aggregator[RequestKey, AccumulatedState, AddressingConfig] {
+    extends GenericAggregator[RequestKey, AccumulatedState, AddressingConfig] {
 
-  def handleUpstreamResponses(
-      handlerInput: HandlerInput
-  ): HandlerResult
+  def handleUpstreamResponses(initialState: AccumulatedState,
+                              upstreamResponseMap: ResponseMap): HandlerResult
 
   def intermediateResponseHandlers: Seq[ResponseHandler] =
     Seq(handleUpstreamResponses)
